+++ src/js/wrapper.ts
@@ -13,6 +13,7 @@ import { NativeModules, Platform } from "react-native";
 import {
   NativeAppStartResponse,
   NativeDeviceContextsResponse,
+  NativeFramesResponse,
   NativeReleaseResponse,
   SentryNativeBridgeModule,
 } from "./definitions";
@@ -45,6 +46,9 @@ interface SentryNativeWrapper {
   fetchNativeRelease(): PromiseLike<NativeReleaseResponse>;
   fetchNativeDeviceContexts(): PromiseLike<NativeDeviceContextsResponse>;
   fetchNativeAppStart(): PromiseLike<NativeAppStartResponse | null>;
+  fetchNativeFrames(): PromiseLike<NativeFramesResponse | null>;
+
+  disableNativeFramesTracking(): void;
 
   addBreadcrumb(breadcrumb: Breadcrumb): void;
   setContext(key: string, context: { [key: string]: unknown } | null): void;
@@ -260,6 +264,17 @@ export const NATIVE: SentryNativeWrapper = {
     return RNSentry.fetchNativeAppStart();
   },
 
+  async fetchNativeFrames(): Promise<NativeFramesResponse | null> {
+    if (!this.enableNative) {
+      throw this._DisabledNativeError;
+    }
+    if (!this._isModuleLoaded(RNSentry)) {
+      throw this._NativeClientError;
+    }
+
+    return RNSentry.fetchNativeFrames();
+  },
+
   /**
    * Triggers a native crash.
    * Use this only for testing purposes.
@@ -420,6 +435,17 @@ export const NATIVE: SentryNativeWrapper = {
     return RNSentry.closeNativeSdk().then(() => {
       this.enableNative = false;
     });
+  },
+
+  disableNativeFramesTracking(): void {
+    if (!this.enableNative) {
+      return;
+    }
+    if (!this._isModuleLoaded(RNSentry)) {
+      return;
+    }
+
+    RNSentry.disableNativeFramesTracking();
   },
 
   isNativeTransportAvailable(): boolean {