+++ test/tracing/reactnativetracing.test.ts
@@ -5,17 +5,12 @@ import { IdleTransaction, Transaction } from "@sentry/tracing";
 import { NativeAppStartResponse } from "../../src/js/definitions";
 import { RoutingInstrumentation } from "../../src/js/tracing/routingInstrumentation";
 
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-function mockFunction<T extends (...args: any[]) => any>(
-  fn: T
-): jest.MockedFunction<T> {
-  return fn as jest.MockedFunction<T>;
-}
-
 jest.mock("../../src/js/wrapper", () => {
   return {
     NATIVE: {
       fetchNativeAppStart: jest.fn(),
+      fetchNativeFrames: jest.fn(() => Promise.resolve()),
+      disableNativeFramesTracking: jest.fn(() => Promise.resolve()),
       enableNative: true,
     },
   };
@@ -52,6 +47,7 @@ const getMockHub = () => {
 import { ReactNativeTracing } from "../../src/js/tracing/reactnativetracing";
 import { getTimeOriginMilliseconds } from "../../src/js/tracing/utils";
 import { NATIVE } from "../../src/js/wrapper";
+import { mockFunction } from "../testutils";
 
 beforeEach(() => {
   NATIVE.enableNative = true;
@@ -66,7 +62,9 @@ describe("ReactNativeTracing", () => {
   describe("App Start", () => {
     describe("Without routing instrumentation", () => {
       it("Starts route transaction (cold)", (done) => {
-        const integration = new ReactNativeTracing();
+        const integration = new ReactNativeTracing({
+          enableNativeFramesTracking: false,
+        });
 
         const timeOriginMilliseconds = Date.now();
         const appStartTimeMilliseconds = timeOriginMilliseconds - 100;
@@ -448,6 +446,26 @@ describe("ReactNativeTracing", () => {
         const transaction = mockHub.getScope()?.getTransaction();
 
         expect(transaction).toBeUndefined();
+
+        done();
+      });
+    });
+  });
+
+  describe("Native Frames", () => {
+    it("Does not initialize native frames instrumentation if flag is false", (done) => {
+      const integration = new ReactNativeTracing({
+        enableNativeFramesTracking: false,
+      });
+      const mockHub = getMockHub();
+      integration.setupOnce(addGlobalEventProcessor, () => mockHub);
+
+      setImmediate(() => {
+        expect(integration.nativeFramesInstrumentation).toBeUndefined();
+        // eslint-disable-next-line @typescript-eslint/unbound-method
+        expect(NATIVE.disableNativeFramesTracking).toBeCalledTimes(1);
+        // eslint-disable-next-line @typescript-eslint/unbound-method
+        expect(NATIVE.fetchNativeFrames).not.toBeCalled();
 
         done();
       });